+++ snmp/datadog_checks/snmp/types.py
@@ -11,4 +11,6 @@ MetricDefinition = TypedDict(
     {'type': Literal['gauge', 'rate', 'counter', 'monotonic_count', 'monotonic_count_and_rate'], 'value': float},
 )
 
-MIBSymbol = NamedTuple('MIBSymbol', [('name', str), ('prefix', Tuple[str, ...])])
+MIBSymbol = NamedTuple('MIBSymbol', [('mib', str), ('symbol', str), ('prefix', Tuple[str, ...])])
+
+OIDMatch = NamedTuple('OIDMatch', [('name', str), ('indexes', Tuple[str, ...])])