+++ snmp/tests/conftest.py
@@ -21,7 +21,6 @@ from .common import (
 FILES = [
     # "https://ddintegrations.blob.core.windows.net/snmp/f5.snmprec",
     "https://ddintegrations.blob.core.windows.net/snmp/3850.snmprec",
-    "https://ddintegrations.blob.core.windows.net/snmp/hpe-proliant.snmprec",
 ]
 
 E2E_METADATA = {