+++ snmp/tests/test_unit.py
@@ -2,6 +2,7 @@
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
 
+import copy
 import os
 import time
 from concurrent import futures
@@ -14,7 +15,8 @@ import yaml
 from datadog_checks.base import ConfigurationError
 from datadog_checks.dev import temp_dir
 from datadog_checks.snmp import SnmpCheck
-from datadog_checks.snmp.config import InstanceConfig, ParsedMetric, ParsedTableMetric
+from datadog_checks.snmp.config import InstanceConfig
+from datadog_checks.snmp.parsing import ParsedSymbolMetric, ParsedTableMetric
 from datadog_checks.snmp.resolver import OIDTrie
 from datadog_checks.snmp.utils import _load_default_profiles, oid_pattern_specificity, recursively_expand_base_profiles
 
@@ -29,12 +31,12 @@ def test_parse_metrics(lcd_mock):
     # type: (Any) -> None
     lcd_mock.configure.return_value = ('addr', None)
 
-    # Unsupported metric
-    metrics = [{"foo": "bar"}]  # type: list
     config = InstanceConfig(
         {"ip_address": "127.0.0.1", "community_string": "public", "metrics": [{"OID": "1.2.3", "name": "foo"}]}
     )
 
+    # Unsupported metric.
+    metrics = [{"foo": "bar"}]  # type: list
     with pytest.raises(Exception):
         config.parse_metrics(metrics)
 
@@ -44,7 +46,7 @@ def test_parse_metrics(lcd_mock):
     assert len(oids) == 1
     assert len(parsed_metrics) == 1
     foo = parsed_metrics[0]
-    assert isinstance(foo, ParsedMetric)
+    assert isinstance(foo, ParsedSymbolMetric)
     assert foo.name == 'foo'
 
     # MIB with no symbol or table
@@ -58,7 +60,7 @@ def test_parse_metrics(lcd_mock):
     assert len(oids) == 1
     assert len(parsed_metrics) == 1
     foo = parsed_metrics[0]
-    assert isinstance(foo, ParsedMetric)
+    assert isinstance(foo, ParsedSymbolMetric)
     assert foo.name == 'foo'
 
     # MIB with table, no symbols
@@ -174,6 +176,16 @@ def test_ignore_ip_addresses():
         SnmpCheck('snmp', {}, [instance])
 
 
+def test_empty_values():
+    instance = common.generate_instance_config(common.SUPPORTED_METRIC_TYPES)
+    instance['user'] = ''
+    instance['enforce_mib_constraints'] = ''
+    instance['timeout'] = ''
+    instance['retries'] = ''
+    check = SnmpCheck('snmp', {}, [instance])
+    assert check._config.enforce_constraints
+
+
 def test_profile_error():
     instance = common.generate_instance_config([])
     instance['profile'] = 'profile1'
@@ -191,6 +203,22 @@ def test_profile_error():
         init_config = {'profiles': {'profile1': {'definition_file': profile_file}}}
         with pytest.raises(ConfigurationError):
             SnmpCheck('snmp', init_config, [instance])
+
+
+def test_duplicate_sysobjectid_error():
+    profile1 = {'sysobjectid': '1.3.6.1.4.1.30932.*'}
+    profile2 = copy.copy(profile1)
+
+    instance = common.generate_instance_config([])
+    init_config = {'profiles': {'profile1': {'definition': profile1}, 'profile2': {'definition': profile2}}}
+
+    with pytest.raises(ConfigurationError) as e:
+        SnmpCheck('snmp', init_config, [instance])
+    assert "has the same sysObjectID" in str(e.value)
+
+    # no errors are raised
+    init_config['profiles']['profile2']['definition']['sysobjectid'] = '1.3.6.2.4.1.30932.*'
+    SnmpCheck('snmp', init_config, [instance])
 
 
 def test_no_address():
@@ -288,6 +316,7 @@ def test_cache_corrupted(write_mock, read_mock):
 @mock.patch("datadog_checks.snmp.snmp.write_persistent_cache")
 def test_cache_building(write_mock, read_mock):
     instance = common.generate_instance_config(common.SUPPORTED_METRIC_TYPES)
+    instance['timeout'] = 1
     instance.pop('ip_address')
 
     read_mock.return_value = '[]'