+++ mesos_master/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG - mesos_master
 
+## 1.7.2 / 2020-04-24
+
+* [Fixed] Handle missing role metrics for mesos_master. See [#6422](https://github.com/DataDog/integrations-core/pull/6422).
+
 ## 1.7.1 / 2020-04-04
 
 * [Fixed] Update deprecated imports. See [#6088](https://github.com/DataDog/integrations-core/pull/6088).